+++ app/src/main/java/com/nullpointer/runningcompose/models/config/MapConfig.kt
@@ -4,5 +4,5 @@ import com.nullpointer.runningcompose.models.types.MapStyle
 
 data class MapConfig(
     val style: MapStyle,
-    val weight:Float
+    val weight:Int
 )
